docs: Add crate-level docs for red_moon
It's perhaps worth discussing if this should be folded into the README,
but the scope of the README likely should be a bit broader to encompass
the `mlua` and `cli` sister crates. Additionally, I did not fully copy
the section listing what is and isn't supported; though this is
certainly very important to include in the crate docs, it seemed tedious
to enforce synchronization between the two locations.